Move product block editor to new feature name
