+++ plugins/woocommerce/client/admin/config/development.json
@@ -2,7 +2,7 @@
 	"features": {
 		"activity-panels": true,
 		"analytics": true,
-		"block-editor-feature-enabled": true,
+		"product-block-editor": true,
 		"coupons": true,
 		"customer-effort-score-tracks": true,
 		"import-products-task": true,
