+++ plugins/woocommerce/src/Admin/Features/Features.php
@@ -40,7 +40,7 @@ class Features {
 	protected static $beta_features = array(
 		'navigation',
 		'new-product-management-experience',
-		'block-editor-feature-enabled',
+		'product-block-editor',
 		'settings',
 	);
